+++ modules/database/src/adapters/sequelize-adapter/postgres-adapter/PostgresSchema.ts
@@ -1,16 +1,9 @@
 import { isNil } from 'lodash';
-import {
-  FindAttributeOptions,
-  FindOptions,
-  ModelStatic,
-  Sequelize,
-  Transaction,
-} from 'sequelize';
-import { MultiDocQuery, ParsedQuery, Query, SingleDocQuery } from '../../../interfaces';
+import { ModelStatic, Sequelize, Transaction } from 'sequelize';
+import { ParsedQuery, SingleDocQuery } from '../../../interfaces';
 import ConduitGrpcSdk, { ConduitSchema, Indexable } from '@conduitplatform/grpc-sdk';
 import { SequelizeSchema } from '../SequelizeSchema';
 import { PostgresAdapter } from './index';
-import { parseQuery } from '../parser';
 
 const incrementDbQueries = () =>
   ConduitGrpcSdk.Metrics?.increment('database_queries_total');
@@ -32,193 +25,6 @@ export class PostgresSchema extends SequelizeSchema {
     },
   ) {
     super(sequelize, schema, originalSchema, adapter as any, extractedRelations);
-  }
-
-  sync() {
-    const syncOptions = { alter: true };
-    let promiseChain: Promise<any> = this.model.sync(syncOptions);
-    incrementDbQueries();
-    return promiseChain;
-  }
-
-  async create(query: SingleDocQuery, transaction?: Transaction) {
-    let parsedQuery: ParsedQuery;
-    if (typeof query === 'string') {
-      parsedQuery = JSON.parse(query);
-    } else {
-      parsedQuery = query;
-    }
-    parsedQuery.createdAt = new Date();
-    parsedQuery.updatedAt = new Date();
-    incrementDbQueries();
-    let relationObjects = this.extractRelationsModification(parsedQuery);
-    let t: Transaction | undefined = transaction;
-    let transactionProvided = transaction !== undefined;
-    if (!transactionProvided) {
-      t = await this.sequelize.transaction({ type: Transaction.TYPES.IMMEDIATE });
-    }
-
-    return await this.model
-      .create(parsedQuery, {
-        transaction: t,
-      })
-      .then(doc => this.createWithPopulation(doc, relationObjects, t))
-      .then(doc => {
-        if (!transactionProvided) {
-          t!.commit();
-        }
-        return doc;
-      })
-      .then(doc => (doc ? doc.toJSON() : doc))
-      .catch(err => {
-        if (!transactionProvided) {
-          t!.rollback();
-        }
-        throw err;
-      });
-  }
-
-  async createMany(query: MultiDocQuery) {
-    let parsedQuery: ParsedQuery[];
-    if (typeof query === 'string') {
-      parsedQuery = JSON.parse(query);
-    } else {
-      parsedQuery = query;
-    }
-    incrementDbQueries();
-    let relationObjects = this.extractManyRelationsModification(parsedQuery);
-    let t = await this.sequelize.transaction({ type: Transaction.TYPES.IMMEDIATE });
-    return this.model
-      .bulkCreate(parsedQuery, {
-        transaction: t,
-      })
-      .then(docs => {
-        return Promise.all(
-          docs.map((doc, index) =>
-            this.createWithPopulation(doc, relationObjects[index], t),
-          ),
-        );
-      })
-      .then(docs => {
-        t.commit();
-        return docs;
-      })
-      .then(docs => (docs ? docs.map(doc => (doc ? doc.toJSON() : doc)) : docs))
-      .catch(err => {
-        t.rollback();
-        throw err;
-      });
-  }
-
-  async findOne(query: Query, select?: string, populate?: string[]) {
-    let parsedQuery: ParsedQuery | ParsedQuery[];
-    if (typeof query === 'string') {
-      parsedQuery = JSON.parse(query);
-    } else {
-      parsedQuery = query;
-    }
-    let parsingResult = parseQuery(parsedQuery, this.extractedRelations, populate);
-
-    const options: FindOptions = {
-      where: parsingResult.query,
-      nest: true,
-      include: this.includeRelations(parsingResult.requiredRelations, populate || []),
-    };
-    options.attributes = {
-      exclude: [...this.excludedFields],
-    } as unknown as FindAttributeOptions;
-    if (!isNil(select) && select !== '') {
-      options.attributes = this.parseSelect(select);
-    }
-    incrementDbQueries();
-    const document = await this.model
-      .findOne(options)
-      .then(doc => (doc ? doc.toJSON() : doc));
-
-    return document;
-  }
-
-  async findMany(
-    query: Query,
-    skip?: number,
-    limit?: number,
-    select?: string,
-    sort?: { [field: string]: -1 | 1 },
-    populate?: string[],
-  ): Promise<any> {
-    let parsedQuery: ParsedQuery | ParsedQuery[];
-    if (typeof query === 'string') {
-      parsedQuery = JSON.parse(query);
-    } else {
-      parsedQuery = query;
-    }
-    let parsingResult = parseQuery(parsedQuery, this.extractedRelations, populate);
-    const options: FindOptions = {
-      where: parsingResult.query,
-      nest: true,
-      include: this.includeRelations(parsingResult.requiredRelations, populate || []),
-    };
-    options.attributes = {
-      exclude: [...this.excludedFields],
-    } as unknown as FindAttributeOptions;
-    if (!isNil(skip)) {
-      options.offset = skip;
-    }
-    if (!isNil(limit)) {
-      options.limit = limit;
-    }
-    if (!isNil(select) && select !== '') {
-      options.attributes = this.parseSelect(select);
-    }
-    if (!isNil(sort)) {
-      options.order = this.parseSort(sort);
-    }
-
-    const documents = await this.model
-      .findAll(options)
-      .then(docs => (docs ? docs.map(doc => (doc ? doc.toJSON() : doc)) : docs));
-
-    return documents;
-  }
-
-  deleteOne(query: Query) {
-    let parsedQuery: ParsedQuery | ParsedQuery[];
-    if (typeof query === 'string') {
-      parsedQuery = JSON.parse(query);
-    } else {
-      parsedQuery = query;
-    }
-    incrementDbQueries();
-    let parsingResult = parseQuery(parsedQuery, this.extractedRelations, undefined);
-    return this.model
-      .findOne({
-        where: parsingResult.query,
-        include: this.includeRelations(parsingResult.requiredRelations, []),
-      })
-      .then(doc => {
-        doc?.destroy();
-      });
-  }
-
-  deleteMany(query: Query) {
-    let parsedQuery: ParsedQuery | ParsedQuery[];
-    if (typeof query === 'string') {
-      parsedQuery = JSON.parse(query);
-    } else {
-      parsedQuery = query;
-    }
-    incrementDbQueries();
-    let parsingResult = parseQuery(parsedQuery, this.extractedRelations);
-    return this.model
-      .findAll({
-        where: parsingResult.query,
-        include: this.includeRelations(parsingResult.requiredRelations, []),
-      })
-      .then(docs => {
-        if (docs) {
-          return Promise.all(docs.map(doc => doc.destroy()));
-        }
-      });
   }
 
   async findByIdAndUpdate(
@@ -325,53 +131,5 @@ export class PostgresSchema extends SequelizeSchema {
       }
       throw e;
     }
-  }
-
-  async updateMany(filterQuery: Query, query: SingleDocQuery, populate?: string[]) {
-    let parsedQuery: ParsedQuery;
-    if (typeof query === 'string') {
-      parsedQuery = JSON.parse(query);
-    } else {
-      parsedQuery = query;
-    }
-    let parsedFilter: ParsedQuery;
-    if (typeof filterQuery === 'string') {
-      parsedFilter = JSON.parse(filterQuery);
-    } else {
-      parsedFilter = filterQuery;
-    }
-    let parsingResult = parseQuery(parsedFilter, this.extractedRelations, populate);
-    incrementDbQueries();
-    let docs = await this.model.findAll({
-      where: parsingResult.query,
-      attributes: ['_id'],
-      include: this.includeRelations(parsingResult.requiredRelations, populate || []),
-    });
-    const t = await this.sequelize.transaction({ type: Transaction.TYPES.IMMEDIATE });
-    try {
-      let data = await Promise.all(
-        docs.map(doc => this.findByIdAndUpdate(doc._id, parsedQuery, populate, t)),
-      );
-      await t.commit();
-      return data;
-    } catch (e) {
-      await t.rollback();
-      throw e;
-    }
-  }
-
-  countDocuments(query: Query): Promise<number> {
-    let parsedQuery: ParsedQuery | ParsedQuery[];
-    if (typeof query === 'string') {
-      parsedQuery = JSON.parse(query);
-    } else {
-      parsedQuery = query;
-    }
-    incrementDbQueries();
-    let parsingResult = parseQuery(parsedQuery, this.extractedRelations);
-    return this.model.count({
-      where: parsingResult.query,
-      include: this.includeRelations(parsingResult.requiredRelations, []),
-    });
   }
 }
